Dictionaries
Print the keys of the boy_names dictionary.
Print the keys of the boy_names dictionary for the year 2013.
Loop over the boy_names dictionary.
Inside the loop, safely print the year and the third ranked name. Print 'Unknown' if the third ranked name is not found.

#####
Excellent work! Dealing with nested dictionaries can be tricky. Here, the 2012 key consists of an empty dictionary, and so there is no third ranked name. As a result, 'Unknown' is printed, unlike for 2013 and 2014.