Create Altering Dictionaries
Assign the names_2011 dictionary as the value to the 2011 key of the boy_names dictionary.
Update the 2012 key in the boy_names dictionary with the following data in a list of tuples: (1, 'Casey'), (2, 'Aiden').
Loop over the boy_names dictionary.
Inside the first for loop, use another for loop to loop over and sort the data for each year of boy_names by descending rank.
Make sure you have a rank and print 'No Data Available' if not. This has been done for you.
Safely print the year and least popular name or 'Not Available' if it is not found. Take advantage of the .get() method.

*******
Great work! It looks like 'Yonah', 'Yitzchak', and 'Yidel' were the least popular baby names in 2011, 2013, and 2014. The dictionary in the 2012 key of boy_names did not contain any data until you added 'Casey' and 'Aiden', so unsurprisingly, 'Aiden' is the least popular name.