Create Checking dictionaries for data
Check to see if 2011 is in the baby_names dictionary.
Print 'Found 2011' if it is present.
Check to see if 1 is in baby_names[2012].
Print 'Found Rank 1 in 2012' if found and 'Rank 1 missing from 2012' if not found.
Check to see if rank 5 is in baby_names[2013].
Print 'Found Rank 5' if it is present.

***************

Superb! If you explore baby_names[2012] in the IPython Shell, you'll see that it is empty. Similarly, baby_names does not contain 2011.