Create Reading from a file using CSV reader
Import the python csv module.
Create a Python file object in read mode for baby_names.csv called csvfile.
Loop over a csv reader on the file object. Inside the loop:
Print each row.

********

Fantastic work! CSV files are among the most common methods of storing tabular data, and you'll encounter them in the wild very frequently. As a result, knowing how to leverage the csv module can be a great addition to your data science toolbox.
Add the rank (the 6th element of row) as the key and name (the 4th element of row) as the value to the existing dictionary (baby_names).
Print the keys of baby_names.

**************