Creating dictionaries of an unknown structure
Create an empty dictionary called ridership.
Iterate over entries, unpacking it into the variables date, stop, and riders.
Check to see if the date already exists in the ridership dictionary. If it does not exist, create an empty list for the date key.
Append a tuple consisting of stop and riders to the date key of the ridership dictionary.
Print the ridership for '03/09/2016'.