Safely appending to a key's value list
Import defaultdict from collections.
Create a defaultdict with a default type of list called ridership.
Iterate over the list entries, unpacking it into the variables date, stop, and riders, exactly as you did in the previous exercise.
Use stop as the key of the ridership dictionary and append riders to its value.
Print the first 10 items of the ridership dictionary. You can use the .items() method for this. Remember, you have to convert ridership.items() to a list before slicing.