Create Working with OrderedDictionaries
Import OrderedDict from collections.
Create an OrderedDict called ridership_date.
Iterate over the list entries, unpacking it into date and riders.
If a key does not exist in ridership_date for the date, set it equal to 0 (if only you could use defaultdict here!)
Add riders to the date key of ridership_date.
Print the first 31 records. Remember to convert the items into a list.