Update Date &Time
*********************

Create a defaultdict of an integer called monthly_total_rides.
Loop over the list daily_summaries, which contains the columns mentioned above in the assignment text.
Convert the service_date (1st element of daily_summary) to a datetime object called service_datetime. Use '%m/%d/%Y' as your format string.
Use the month of the service_datetime as the dict key and add the total_rides (5th element of daily_summary) to the current amount for the month. Be sure to convert this into an integer.
Print monthly_total_rides.